repository

+++ EmployeeManagementSystem/src/main/java/com/employee/model/employee.java
@@ -13,15 +13,16 @@ public class employee {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long id;
+	private Integer id;
 	private String firstame;
 	private String lastame;
 	private String email;
 	
-	public long getId() {
+	
+	public Integer getId() {
 		return id;
 	}
-	public void setId(long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 	public String getFirstame() {
